Venn: - added venn diagram where the circles are shaded by a gradient

+++ VennDiagrams/Venn/src/venn/BallVennDiagram.java
@@ -151,7 +151,7 @@ public class BallVennDiagram extends VennDiagramTemplate {
 
 		for(int i : getPartialShapeIndices()) {
 			Shape shape = getPartialShape(i);
-			int size = getData().getUnionCount(i);
+			double size = getData().getUnionCount(i);
 
 			double ratio = size / getMaxBallCount(shape);
 			maxRatio = Math.max(maxRatio, ratio);
